# docs/openapi.yaml
openapi: 3.0.3
info:
  title: The Ultimate API
  description: |
    # 💀 The Ultimate API 
    
    Welcome to **The Ultimate API** - the most comprehensive and powerful Roblox audio downloading service available. This API provides seamless access to Roblox's vast audio library with advanced features including batch processing, detailed analytics, user management, and robust rate limiting.
    
    ## 🚀 Key Features
    
    - **🎵 Audio Downloads**: Download any public Roblox audio file instantly
    - **📦 Batch Processing**: Download multiple audio files in a single request
    - **👥 User Management**: Secure authentication with JWT tokens and API keys
    - **📊 Advanced Analytics**: Detailed statistics and usage tracking
    - **🛡️ Rate Limiting**: Smart rate limiting to prevent abuse
    - **⚡ High Performance**: Async processing for maximum speed
    - **🔒 Secure**: Industry-standard security practices
    
    ## 🛠️ Getting Started
    
    1. **Register** a new account using the `/auth/register` endpoint
    2. **Login** to get your JWT token via `/auth/login`
    3. **Generate** an API key (optional) via `/auth/api-key`
    4. **Start downloading** audio files using `/audio/download`
    
    ## 🔐 Authentication
    
    This API supports two authentication methods:
    - **JWT Tokens**: Perfect for web applications and temporary access
    - **API Keys**: Ideal for server-to-server communication and long-term access
    
    ## 📈 Rate Limits
    
...
    
...
    
    ```
...
servers:
...
paths:
...
  /audio/download:
    post:
      tags:
        - Audio
      summary: Download Audio File
      description: |
        Download a single Roblox audio file by its asset ID.
        
        **Features:**
        - Lightning-fast downloads
        - Automatic file format detection
        - Metadata extraction
        - Error handling and retry logic
        
        **Process:**
        1. Validates the Roblox asset ID
        2. Fetches audio metadata from Roblox
        3. Downloads the audio file
        4. Returns download information
        5. Logs the download for statistics
        
        **Supported Formats:**
        - MP3 (most common)
        - OGG
        - WAV
        - M4A
        
        **File Information Provided:**
        - Original filename
        - File size
        - Duration (if available)
        - Audio quality metrics
        - Download speed statistics
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AudioDownloadRequest'
            examples:
              single_download:
                summary: Download Popular Audio
                value:
                  asset_id: 142376088
                  include_metadata: true
      responses:
        '200':
          description: Audio downloaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudioDownloadResponse'
              examples:
                download_success:
                  summary: Successful Download
                  value:
                    asset_id: 142376088
                    filename: "Roblox_Oof_Sound.mp3"
                    file_size: 24576
                    download_url: "https://api.ultimate-roblox.com/files/temp/abc123.mp3"
                    duration: 1.2
                    format: "mp3"
                    quality: "128kbps"
                    success: true
                    download_time: 0.85
                    metadata:
                      title: "Oof Sound Effect"
                      creator: "Roblox"
                      created_date: "2013-04-01T00:00:00Z"
        '400':
          description: Invalid asset ID or download failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid_asset:
                  summary: Invalid Asset ID
                  value:
                    detail: "Asset ID 999999999 not found or not accessible"
                    error_code: "ASSET_NOT_FOUND"
                download_failed:
                  summary: Download Failed
                  value:
                    detail: "Failed to download audio file"
                    error_code: "DOWNLOAD_ERROR"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /audio/batch:
    post:
      tags:
        - Audio
      summary: Batch Download Audio Files
      description: |
        Download multiple Roblox audio files in a single request.
        
        **Batch Processing Features:**
        - Concurrent downloads for maximum speed
        - Individual file status tracking
        - Partial success handling
        - Progress monitoring
        - Automatic retry for failed downloads
        
        **Limits:**
        - Maximum 50 files per batch request
        - Total size limit: 500MB per batch
        - Processing timeout: 5 minutes
        
        **Response Details:**
        - Individual success/failure status for each file
        - Download statistics and timing
        - Error details for failed downloads
        - ZIP archive option for multiple files
        
        **Best Practices:**
        - Use smaller batches for better performance
        - Monitor the response for partial failures
        - Implement retry logic for failed items
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchDownloadRequest'
            examples:
              batch_download:
                summary: Download Multiple Audio Files
                value:
                  asset_ids: [142376088, 131961136, 130768997]
                  include_metadata: true
                  create_zip: false
      responses:
        '200':
          description: Batch download completed (may include partial failures)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchDownloadResponse'
              examples:
                batch_success:
                  summary: Successful Batch Download
                  value:
                    total_requested: 3
                    successful_downloads: 3
                    failed_downloads: 0
                    total_size: 73728
                    processing_time: 2.45
                    downloads:
                      - asset_id: 142376088
                        filename: "Oof_Sound.mp3"
                        success: true
                        file_size: 24576
                        download_time: 0.85
                      - asset_id: 131961136
                        filename: "Click_Sound.mp3"
                        success: true
                        file_size: 12288
                        download_time: 0.45
                      - asset_id: 130768997
                        filename: "Victory_Sound.mp3"
                        success: true
                        file_size: 36864
                        download_time: 1.15
        '400':
          description: Invalid request or too many files
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token authentication. Include the token in the Authorization header:
        ```
        Authorization: Bearer <your_jwt_token>
        ```
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication. Include your API key in the X-API-Key header:
        ```
        X-API-Key: uk_live_your_api_key_here
        ```
  schemas:
...
    AudioDownloadRequest:
      type: object
      required:
        - asset_id
      properties:
        asset_id:
          type: integer
          description: Roblox asset ID of the audio file
          example: 142376088
        include_metadata:
          type: boolean
          description: Include audio metadata in response
          default: true
          example: true
    AudioDownloadResponse:
      type: object
      properties:
        asset_id:
          type: integer
          description: Roblox asset ID
          example: 142376088
        filename:
          type: string
          description: Downloaded filename
          example: "Roblox_Oof_Sound.mp3"
        file_size:
          type: integer
          description: File size in bytes
          example: 24576
        download_url:
          type: string
          description: Temporary download URL
          example: "https://api.ultimate-roblox.com/files/temp/abc123.mp3"
        duration:
          type: number
          description: Audio duration in seconds
          example: 1.2
        format:
          type: string
          description: Audio format
          example: "mp3"
        quality:
          type: string
          description: Audio quality
          example: "128kbps"
        success:
          type: boolean
          description: Download success status
          example: true
        download_time:
          type: number
          description: Download time in seconds
          example: 0.85
        metadata:
          type: object
          description: Audio metadata
          properties:
            title:
              type: string
              example: "Oof Sound Effect"
            creator:
              type: string
              example: "Roblox"
            created_date:
              type: string
              format: date-time
              example: "2013-04-01T00:00:00Z"
    BatchDownloadRequest:
      type: object
      required:
        - asset_ids
      properties:
        asset_ids:
          type: array
          items:
            type: integer
          maxItems: 50
          description: Array of Roblox asset IDs (max 50)
          example: [142376088, 131961136, 130768997]
        include_metadata:
          type: boolean
          description: Include metadata for each file
          default: true
          example: true
        create_zip:
          type: boolean
          description: Create ZIP archive of all files
          default: false
          example: false
    BatchDownloadResponse:
      type: object
      properties:
        total_requested:
          type: integer
          description: Total files requested
          example: 3
        successful_downloads:
          type: integer
          description: Number of successful downloads
          example: 3
        failed_downloads:
          type: integer
          description: Number of failed downloads
          example: 0
        total_size:
          type: integer
          description: Total size of all files in bytes
          example: 73728
        processing_time:
          type: number
          description: Total processing time in seconds
          example: 2.45
        downloads:
          type: array
          items:
            $ref: '#/components/schemas/AudioDownloadResponse'
...
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error description
          example: "Asset ID not found"
        error_code:
          type: string
          description: Machine-readable error code
          example: "ASSET_NOT_FOUND"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2025-07-13T12:30:00Z"